backend/requirements.txt
fastapi
uvicorn
pandas
numpy
psycopg2-binary
python-dotenv
requests
scikit-learn
joblib
paho-mqtt
supabase
twilio
apscheduler
google-generativeai
pytz
python-dateutil
